Splitted BSP server implementation
